feat(deploy-aws): use different webpack config for dev and prod environment
