+++ packages/graphql/webpack.common.js
@@ -1,0 +1,55 @@
+const path = require('path');
+
+const webpack = require('webpack');
+const TerserPlugin = require('terser-webpack-plugin');
+const WebpackNodeExternals = require('webpack-node-externals');
+
+module.exports = {
+  target: 'node',
+  optimization: {
+    minimizer: [new TerserPlugin()],
+  },
+  externals: [
+    WebpackNodeExternals({
+      whitelist: ['webpack/hot/poll?1000'],
+    }),
+    WebpackNodeExternals({
+      modulesDir: path.resolve(__dirname, '../../node_modules'),
+      whitelist: [/@animavita/],
+    }),
+    {'aws-sdk': 'aws-sdk'},
+  ],
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js', '.json', '.mjs'],
+  },
+  output: {
+    path: path.resolve(__dirname, 'build'),
+    filename: 'index.js',
+    library: 'index',
+    libraryTarget: 'commonjs2',
+  },
+  plugins: [
+    new webpack.DefinePlugin({
+      'process.env.BROWSER': false,
+      __DEV__: process.env.NODE_ENV !== 'production',
+    }),
+  ],
+  module: {
+    rules: [
+      {
+        test: /\.mjs$/,
+        include: /node_modules/,
+        type: 'javascript/auto',
+      },
+      {
+        test: /\.([jt]sx?)$/,
+        loader: 'babel-loader',
+        options: {
+          configFile: './babel.config.js',
+        },
+        exclude: /node_modules/,
+        include: [path.resolve(__dirname, 'src'), path.resolve(__dirname, '../')],
+      },
+    ],
+  },
+};
