+++ packages/graphql/webpack.prod.js
@@ -1,0 +1,9 @@
+const webpack = require('webpack');
+const merge = require('webpack-merge');
+
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  mode: 'production',
+  entry: ['./src/index.ts'],
+});
